+++ src/main/java/com/artemget/oil_service/repository/RecordRepositoryImpl.java
@@ -23,7 +23,7 @@ public class RecordRepositoryImpl implements RecordRepository {
     }
 
     @Override
-    public void removeRecord(int recordId) {
+    public void removeRecord(long recordId) {
         recordDataSource.deleteRecordById(recordId);
     }
